ConsoleExample builds on mac, but is missing: - File notification - Runtime compilation - Some thread safety code (and the threads) - Window management code (the rocketlibsystem stuff) - TimeSystem does nowt Also: - It needs testing on windows to check it still works - I've broken the friend/private behaviour of a couple of classes for gcc

+++ Aurora/Systems/RocketLibSystem/RocketLibSystemRenderInterfaceOpenGL.cpp
@@ -51,7 +51,7 @@ void RocketLibSystemRenderInterfaceOpenGL::RenderGeometry(Rocket::Core::Vertex* 
 	else
 	{
 		glEnable(GL_TEXTURE_2D);
-		glBindTexture(GL_TEXTURE_2D, (GLuint) texture);
+		glBindTexture(GL_TEXTURE_2D, (intptr_t) texture);
 		glEnableClientState(GL_TEXTURE_COORD_ARRAY);
 		glTexCoordPointer(2, GL_FLOAT, sizeof(Rocket::Core::Vertex), &vertices[0].tex_coord);
 	}
